Move GuildCache to guild.rs
