+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod client;
 pub mod commands;
+pub mod guild;
 pub mod handlers;
-pub mod settings;
 pub mod strings;
 pub mod tests;
 pub mod utils;
